[graphite] Support one-time-only tasks in Recordings
This allows Tasks to report (success, discard, fail) from both
prepareResources() and addCommands(). Success was the previous
'true' value, Fail matches 'false'. Discard is another "success"
that tells the caller to discard the task afterwards. Specifically,
this discards the reference to that task; if the task is referenced
multiple times throughout the task list/graph, those other references
will have prepareResources() and addCommands() called and can choose
to discard or return success/fail.

If prepareResources() discards the task, addCommands() should not be
called for it. If addCommands() returns discard, a replayed Recording
will skip the task.

At the moment, only UploadTasks take advantage of it because of the
OneTimeUploadContext. In the future, this can hopefully be used to
wrap one-time mipmap generation for textures that require it, or for
copies that should not be repeated.

In the near term, this will be used with DrawTasks that get added to
a task list multiple times (initially always to the root task list,
but that will go away, and then any time the scratch image is
drawn or copied from). The extra references to the same draw task
will return discard after performing read/write tracking in
prepareResources() so that after that phase is finished, there will
only be the first reference to the task to perform addCommands().

Bug: b/333052502
Change-Id: I2f1d449b0ecf18391f89e7406bf4d2ba7a0da308
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/836136
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>