[graphite] Add RenderPassDesc to pipeline labels and traces
Replaces AttachmentTypes.h with RenderPassDesc.h/cpp and moves the
LoadOp and StoreOp enums to ResourceTypes.h.

Moves RenderPassDesc::Make() definition out of DrawContext.cpp and into
RenderPassDesc.cpp.

Adds RenderPassDesc::toString() based on what had been in
UniqueKeyUtils.

Updates ContextUtils to take the RenderPassDesc in fully instead of
just the write swizzle so that it can include in the generated label.

Updates the trace event in createGraphicsPipeline() to include the
RenderPassDesc as well.

Change-Id: Ie74ce00f9364d0efbef2916b2c92025a10e02867
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/845338
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>