Add a FixedArray class to shrink SkSL::ComponentArray.
Using a fixed-size array here saves about 2K of code size, and uses
about 16K less RAM when loading the Graphite modules.

The FixedArray class was designed to be a drop-in replacement for
STArray (at least, enough to pass the basic tests in TArrayTest).
For now, it only supports trivial types and up to 255 elements, but
these restrictions are fixable if a need arises.

Change-Id: I09358427a144af24deb2031813f95759f48efd1f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/844437
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>