Revert "[graphite] Revert KeyContext copy ctor behavior to earlier state"
This change does seem to be the cause for most of the perf regressions found on:

https://skia-review.googlesource.com/c/skia/+/842376 ([graphite] Reduce shader variation due to LocalMatrixShaders)

That said, the change to KeyContext's copy ctor is definitely more correct (So I'm reverting this experiment).

What is going on is that, without the bug fix, the KeyContext's kRuntimeEffect scope is getting reset to kDefault (usually in KeyContextWithLocalMatrix). This means that in SkImageShader's add_to_key method we then add an extra blend w/ the Paint's RGB for
alpha-only images that should be being treated as raw (as a child of a blur known-runtime-effect). This is a bit counter intuitive since the bug was adding more work to the program not less.

Change-Id: I49d86061876ac39ff119897a07e304486e3f50e1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/847656
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>