[graphite] Reduce the SkMatrixConvolutionImageFilter's program combinations
This CL switches the MatrixConvolutionImageFilter to only support two sizes in its texture-based code path. This yields 3 shading programs:

<= 28  uniform-based
<= 64  medium texture-based
<= 256 max texture-based (and max allowed)

The CL:

https://skia-review.googlesource.com/c/skia/+/823036 (Reduce matrix convolution image filter's maximum kernel size)

made it into Chrome's M124 branch which has been in Stable release since 4/17. The lack of bug reports indicates we can proceed w/ simplifying the MatrixConvolutionImageFilter. This CL is setting up to convert the MatrixConvolutionImageFilter to be stably keyed.

Performance-wise, on an M1 mac for the matrixconvolution GMs, the uniform-based and 7x7 are basically the same performance before and after this CL. The 128x1 case is 12% slower on Ganesh mtl and ~30% slower on Graphite grmtl. The 1x255 case is ~10% slower for both Ganesh and Graphite.

For Graphite the numbers don't appreciably change if we add a 128 special case. For Ganesh, the 128x1 case improves to be only 2% slower.

Of note, Graphite's performance on the texture-based matrix convolution draws is, across the board, notably slower than Ganesh's.

Bug: b/238759147
Change-Id: I582a484bdbd1b7ca1b336c6f3663537859dfe0a3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/847737
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>