[graphite] Call notifyInUse() on images not known until AddToKey is called
If a client image provider called TextureFromImage on a picture-backed
image that got to AddToKey inside an SkShader, the returned graphite
image would have a linked Device from when the picture was rendered
into its surface. Without the notify call, and if the image provider
cached the images, the Device and DrawContext would not be cleaned up
until the image was removed from the client's cache.

A DrawList is 8kb so if the client cached lots of small images, the
leaked devices and contexts could add up to a meaningful number of MB.

Bug: b/338453542
Change-Id: I30845999e68ace6e969b3ff91952944216b36e3f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/849476
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>