[graphite] Don't rewrap TextureProxy in a new Image in RescaleImage()
RescaleImage() had been inspecting the texture proxy, then using that
to create a new Image that was effectively the same as `srcImage`,
except that now that Devices are linked to dynamic images, it was
losing that linkage when `tempInput` was drawn instead of `srcImage`.

This doesn't really matter while tasks are pushing directly to the
root task list because any dynamic `srcImage` has flushed itself
entirely before calling RescaleImage(). But eventually this needs to
be drawing the `srcImage` so that any dynamic DrawTasks are linked
correctly to the DrawContexts that are downsampling them.

Bug: b/336779772
Change-Id: I0087c98e54adeb92af352f152d6be4415ab2893c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/849822
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>