[graphite] Let Surface automatically generate mipmaps when rendering a picture image
Surface/Device now automatically add a call to GenerateMipmaps if the
target is mipmapped, so this was adding a second set of commands to
redo the work (I think this was here first, but was missed in cleanup
once the auto-generate system was put in).

Also switched to calling Surface::Make() instead of ::RenderTarget to
be able to pass in Budgeted::kNo to make it clear that that is the
behavior we want (compared to when we render a picture for an
SkPictureShader dynamically).

Bug: b/336779772
Change-Id: Ifae395b04382bd6924b775d895c7e38bdc2e20e5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/849896
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>