Update rules to enforce IWYU in the modular build
This adds a target //tools:full_build that has all the modules
as of writing and builds .h.cpp files for our standalone headers
such that IWYU actually analyzes them. This target is then built
with a (much reduced) enforce_iwyu config which uses the existing
iwyu feature [4] that exists in our Linux toolchain.

"Why all the SkDebug.h additions?" In the non-modular build,
we set SK_DEBUG via defines [1] which IWYU is smart enough to
not require a #include for. However, in the modular build, we
don't tack on these defines anymore; defines that change behavior
are now to be set in SkUserConfig.h [2] and defines that advertise
behavior (such as SK_FONTMGR_FONTCONFIG_AVAILABLE) are set on the
skia_cc_library targets explicitly.

That means in order for a file to use SK_DEBUG, it needs to
depend on a file that sets it. Technically this would be
SkLoadUserConfig.h however, we made SkDebug.h pretend to export
this [3] because it's more intuitive.

Suggested review order:
 - //tools/BUILD.bazel
 - bazel/generate_cpp_files_for_headers.bzl to see a small
   cleanup in how we generate .cpp files from standalone headers
   (don't need to list the file twice).
 - All BUILD.bazel changes in //include/... and //src/... The
   list of generated files was broken up from the top level
   BUILD.bazel and moved down into the owning files because this
   could be done without exporting the files (and is a touch cleaner).
 - //modules/skottie/BUILD.bazel where cc_binary_with_flags targets
   were replaced with skia_cc_binary ones depending on the modular
   build.
 - infra/bots/jobs.json to see the changed jobs. I got rid of the
   //src/base job in favor of //:core since the latter is what clients
   would be building/depending on (and is actually implemented now).
 - All .h and .cpp changes (should be mechanical; see above for rationale)
 - All other changes.

[1] https://github.com/google/skia/blob/1a4531b09eeb499915c0c6898eb757ce4b595fed/defines.bzl#L11-L20
[2] Exceptions to this being GR_TEST_UTILS and defines to disable legacy
    codepaths that are not Blazel friendly, e.g. SK_DISABLE_LEGACY_INIT_DECODERS
[3] https://github.com/google/skia/blob/0015580cb26178cb09c832106508c41fa8570e2f/toolchain/linux_trampolines/IWYU_mapping.imp#L445-L447
[4] http://review.skia.org/519776

Bug: b/40044159
Change-Id: I5c319384c455264d9dde5af107ec15a1da01ee16
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/850077
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Owners-Override: Kaylee Lubick <kjlubick@google.com>