[bazel] Add some missing dependencies/files to Ganesh Metal Backend
Changes to GrMtlFramebuffer are to keep GrMtlAttachment as a forward
declare and not a full #include.

See also http://cl/632243533

(How did these slip through? When doing local development, I was
bypassing the sandbox, but didn't actually check that it worked
with the sandbox [sandboxing on Macs is quite slow]. A G3 client
tested it with the sandbox and these problems emerged).

Change-Id: Ib93725c79a51bc6cfa45ea3bdc444a931dbff957
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/851155
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Owners-Override: Kaylee Lubick <kjlubick@google.com>