Change CtsEnforcement for ImageAsyncReadPixels
Follow-on to I4283d79a53a56b4e0aabcf13f4b562bbedccf939. As with
SurfaceAsyncReadPixels, this test was also modified in
I534fbdd622ca11beb192d84107cbcc3629210945. It is effectively a new test
in V, so we can only start enforcing it for devices launching with V and
above, in accordance with GRF.

Bug: 336594045
Change-Id: I297ccf2fac5c0ee256cc5ea9a98e20b7ee480f20
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/853098
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>
Reviewed-by: Nolan Scobie <nscobie@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>