Disable the AAA "deferred" blitting mode
This has different performance characteristics (and different visual
results, in some cases). However, it's additional complexity that we
don't want to maintain, and the perf benefits are minor.

The complexity is a problem: I've identified a bug in the core of the
AAA edge walker, but fixing it in deferred mode is much trickier than
in the default mode. At this point, time to delete it.

Bug: b/341966153
Bug: b/340982297
Change-Id: I19a9c90b2ee4d67388a0fe3329999b17f419b7e8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/855820
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>