[Ganesh] Fix readPixels for planar textures on Vulkan.
When the underlying format is planar,
GrVkCaps::surfaceSupportsReadPixels returns kCopyToTexture2D.  Because
the color type is kRGB_888x and kRGB_888x is not renderable on Vulkan,
the allocation of the temporary surface fails.

When we know the allocation will fail, fall back to kRGBA_8888 instead.

Bug: b/339242357
Change-Id: Ia5b3324dab4794bd2513a006a2aea157d5baf5ee
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/856296
Commit-Queue: Chia-I Wu <olv@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>