Fix scan converter bug causing low coverage slivers in strokes
When inserting new edges, we need to perform intersection tests against
the next edge (not just the previous edge). During the regular sweep of
the edges, we only need to do a backward intersection test (because we
eventually visit all edges, so each interval is checked for
intersection). But the new edge was never compared to it's right
neighbor... So we do that explicitly.

For new edges that will immediately cross an edge to their right, this
increases the scanning resolution, so we don't accidentally use the
wrong left edge over a larger Y range than we should.

Originally discovered via coverage gaps from stroker-generated bow-tie
geometry. The new GM includes two different instances of this, both are
fixed by the new code.

Guarded (using the same define as the deferred blitting) so we can
rebaseline chromium all at once for the changes this will trigger.

Bug: b/340982297
Change-Id: I4229c4bc896831dfbeb5150e92ecbb8fbf419020
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/856596
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>