[graphite] Settle on BlurMaskFilter Precompilation API
At least for the BlurMaskFilter the Precompilation API continues to parallel the main Skia API. This is a bit odd since we allocate an object just to use it as a signal to build the BlurImageFilter pipelines. When (far in the future) we can reduce the main Skia API's blurMask API to just be fields on SkPaint, this overhead can go away.

Bug: b/259548724
Change-Id: I0b46ba083aea296480dbfc865eff8cbced2041e8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/857536
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>