Don't draw paths that are too large to tesselate
If the path's viewport is too large, we currently hit an assert.
This change makes it so that if the viewport is too large, we don't
draw the path.

This also adds early out in chop pathing for a couple of cases that
where chopping the path won't do anything.

Bug: 341126039
Change-Id: If43e97ed887021f158d62a69d497b7c0372860fd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/856216
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>