Cleanup vulkan memory allocator and include in Bazel build (for now)
Ideally, we'd have clients pass in a memory allocator via the
GrVkBackendContext, but we aren't quite ready for that (we'd need
to make a module for the allocator). This cleans up a few things
in the C++ code and the Bazel rules such at our G3 clients
can use the allocator as is.

Bug: b/40045064
Change-Id: I7e65352dd6a954b19a7c620e284cdf76e92a93d4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/857421
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>