[Graphite] Add descriptor set cache for Vulkan uniforms.
This adds a cache of written descriptor sets used for binding
vulkan uniform buffers. The cache key is based off keying each
binding of the set. A binding is keyed using the buffers uniqueID
and the size/range of the binding.

To allow for reuse of these descriptor sets, we now always use
dynamic uniform descriptors. To do this we always write the
descriptor with an offset of 0 and then set the correct offset
when we bind the set.

A future optimization to this caching is to track the last
bound desc set on the command buffer to be able to quickly check
if that desc set can be used again without needing to generate
a key and do the cache look up. Running through the skps
about 50% of the time we get a desc set from the cache it is the
same as the previously used one.


Change-Id: I5d0cf8b25bea4f3b87d6d93a4fbd41b7fec975af
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/854437
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>