Fix text 565 to RGBA conversion on BGRA-based systems.
On systems where a 565 format isn't supported, we convert to RGBA and
use that instead. However, if the base color type for the system is
BGRA, we assume the font data is in that format as well, and
swizzle it before copying to an RGBA texture. The end result is that
we swap R and B, ending up with the incorrect LCD glyph. The solution
is to convert 565 data to BGRA data on those systems.

Bug: b/337905340
Change-Id: I2b005f977a8af7a3de99d93152c615e595daf4f4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/859118
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>