[skif] Disable padding on dithered FilterResults
Chromium still has SK_DONT_PAD_LAYER_IMAGES defined, due to WPT ref
test failures where the dither origin shifts between a regular draw
of a gradient shader and one where it's embedded in an image filter.

It also requires additional regular layout test rebaselines to
deal with interpolation/rendering differences for cases where there is
padding. This is manageable but the WPT ref tests require changes to
behavior to be able to remove the #define (since implementing dither
origin control is non-trivial).

When SK_DONT_PAD_LAYER_IMAGES is defined, the AutoSurface constructor
always uses PixelBoundary::kUnknown regardless of what was passed in.
These changes modify the behavior for when SK_DONT_PAD_LAYER_IMAGES
is not defined and makes it so that when dithering would be used in
an image-filtering context we don't pad the offscreen surfaces.

PS1 tested the change against chrome by removing all pound defines and
the WPT tests passed, so once this change lands, it should be possible
to stage the removal of SK_DONT_PAD_LAYER_IMAGES in chrome.

Bug: b/329700315
Change-Id: I294b1d493c16440613a5c71aaf7abf31ae90513e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/859376
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>