Add uvecN type to the built-in types.
Previously, we did not mention this type anywhere in SkSL. Even
though we don't support it in public Runtime Effects (as unsigned
types don't exist in ES2), it's important to include it so that
the compiler is aware that the type exists and disallows it in
user code. It now works properly in #version 300 as well.

Change-Id: I0c17d1ceb3f6ae78ce46ad2dce907d2b00440f79
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/859061
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>