SkExif: Allow truncated input
Many Exif parsers allow truncated output. That is, they will parse as
much of the Exif that they can read until they hit the end of the
metadata buffer, even if the metadata indicated that the buffer should
have been longer.

This was discovered when migrating Chromium to use the new parser, and
WPT failures caught the change. The parser in blink has this property.
Upon excavation, the parser previously used by Skia did as well (so
this is restoring previously existing behavior).

Update SkTiffImageFileDirectory::MakeFromOffset to have an
allowTruncated parameter, and modify the SkExif parser to specify it.

Also, update SkExif the use std::optional rather than a separate "is
present" variable, since that makes things much shorter.

Add tests to ensure that this doesn't regress.

Bug: b/40283599
Change-Id: I048fd6c3f4acb81d70cc53fc58020d5e7a48a734
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/858616
Reviewed-by: Leon Scroggins <scroggo@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>