[skif] Clone Graphite's shader blur algorithm into src/core as SkShaderBlurAlgorithm
The algorithm independent blur utilities are now defined in
SkBlurEngine, and the blur runtimeeffect related utilities are static
functions in SkShaderBlurAlgorithm. Eventually they will become private
but for now they are visible so that skgpu::BlurUtils can wrap them to
minimize the migration work.

The SkShaderBlurAlgorithm is identical to Graphite's shader algorithm
except that it doesn't have any of the internal rescaling logic
(it will solely be used with FilterResult::rescale()). It also
is virtual to allow each backend to define how to create SkDevices,
so as the backends move to the new system, Graphite's version will
be removed, GrBlurUtils will be removed, and raster will switch to
using this when the color type or tile mode requires it.

Bug: b/294102201
Change-Id: I88b618f775dc7cf5ebccffeaf4df60c6deeb7208
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/861237
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>