[skif] Split blur filling into strict and fast regions
GrBlurUtils already does this (always did this) for clamp and decal,
but not mirror/repeat. Graphite previously was always using a single
strict-shaded draw to fill the blur. For blurs with lots of rescaling,
the final pixel area wasn't that large so it probably didn't matter.
But for large source images with a relatively small blur, the
interior is significantly large enough to make switching to a simpler
shader worthwhile.

Some perf numbers for `GM_animated-tiled-image-blur` on `grmtl` on
my Mac mini:
sigma = 5
 w/ splitting = 1.38ms frame time
 w/o splitting = 2.33ms

sigma = 0.4
 w/ splitting =  1.33ms frame time
 w/o splitting = 2.93ms

Change-Id: I4bf0e59fc8f03361c55b703fd99c75139c007181
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/862998
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>