SkJpegGainmapEncoder::MakeMPF: Make function more readable
The function SkJpegGainmapEncoder::MakeMPF needs to know the offset
into the first image where it will insert its metadata.

In the previous writing of the function, it would determine this
offset while writing the file, creating a "write the file segment
by segment" structure, where at some point in the middle, it would
populate the SkJpegMultiPictureParameters structure.

Change the function to instead compute the offset ahead of time, and
to also only scan the header of the file (up to the StartOfScan
marker).

Add the helper function SkJpegMultiPictureParameters::
GetImageDataOffset to move the offset math from the file. Also rename
GetAbsoluteOffset to GetImageAbsoluteOffset to be clear, and share
the math with GetImageDataOffset.

Bug: b/338342146
Change-Id: I2c134b516820350b84be299333ee302a1aa2d348
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/859776
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>