Remove SkScalerContextRec::fPaintGamma
When the mask gamma hack was first written there was still some debate
about how color spaces would be handled. Skia landed on always
converting the source colors into the destination space before use. As a
result there is no longer any need to handle the case where the paint
color and the device color are in different color spaces, as they are
now always in device space by this point.

This should make no practical difference as the paint and device gamma
always needed to be the same in practice, otherwise superluminous and
subluminous masks would be created and drawing would break in
interesting ways.

Change-Id: I30c976b2018437ed01ab0947f8564452adc72094
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/865039
Reviewed-by: Brian Osman <brianosman@google.com>