[skif] Use FilterResult::rescale() for blur()
Updates blur() to call rescale() if the algorithm's max sigma is too
small so that blur engines do not need to internally provide rescaling.
Improves the intermediate dst bounds calculations in the core shader
blur algorithm.

Adds a staging virtual to blur engine to let backends choose to use
the new blur() that calls rescale() or the old version that calls
resolve(). It turns out that it can't just be staged via reporting
infinity for the max blur sigma because the "identity" behavior when
there are effects in the new rescale() path is just enough different
to trigger layout test failures in chrome, relative to resolve().

This adds logic to Graphite's blur engine to use the core shader
algorithm and opt into the new blur behind a staging flag. Skia
will default to this, but chrome will still use the old graphite
behavior. For now ganesh and raster will remain on legacy and
follow-up CLs will stage them similarly and move their default
behavior to the new system.

Bug: b/294102201
Change-Id: Id9ba23e14b2de4ab8de99892a0c025dfb3cee4f6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/861177
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>