[graphite] Consolidate built-in shader-snippet tables
The major changes are adding default parameters to ShaderSnippet and
allowing it to own the uniform description, so that we can just inline
the initializer lists when filling out the table. The name and
function constants were removed so it was clearer when reading the ID
entry to determine what sksl function to look for in the .frag file.

Part of the default parameters was to allow the preamble and
expression function to be null, which then implicitly uses the
default generation logic. The default generator functions were moved
to be closer to the rest of the sksl generation logic. Functions that
override the default generators are allowed to pass null for the
static function name. The generated expressions are now consistently
using the mangled `fName` and `fStaticFunctionName` is only used when
invoking the module function.

Static asserts for some shader types were moved to the bottom since
they were otherwise now isolated from the rest of the snippet
definition.

While a very large CL, this shouldn't actually be changing the sksl
that is produced for a given paint params key, nor does it change
how the paint param keys need to be structured for a paint.

Change-Id: I9870d104fc810b89c91e3e00244e73dbb316015a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/869038
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>