[graphite] Use sk_FragCoord for dithering
Local coordinates would give incorrect dithering if there was a big
scale factor or orientation relative to the pixel grid. This also
allows the dither snippet to be a "color filter", whereas previously
it was the only snippet that took both a color and a coordinate. This
will simplify some of the expression generation details to remove
that corner case.

Change-Id: Ibbd14b0288aaf152123ed8e7da35aefd6c94c6f1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/869616
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>