[Graphite] Consolidate ssbo caps
Since storage buffers are always preferred when they are supported, it makes sense to consolidate the two caps. That way this will allow for SkSL to know when a shader is using all storage buffers or all uniform buffers when deciding on memory layouts.

Change-Id: I177002c42e5f335ed21dbeeaa456e18189b98857
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/870698
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>