[Graphite] Make gradient buffer support more defined
Currently the gradient buffer approach for large gradients only tightly packs the float data, and it would not be optimal to use std140 since 3/4ths of the bytes would be padding for alignment. This makes it defined that using gradient buffers require a tightly packed layout for float arrays.

Change-Id: I2532c652bb6cc05651fc4558c05ac4d5fd10a72d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/870719
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>