[graphite] Consolidate expression/preamble generators
The compose preamble generator is generalized to invoke the first
N-1 children and then set them as the ShaderSnippet::Args passed to
the Nth child. This lets both the regular compose and the blend
compose use the same generator.

The CoordClamp and LocalMatrix preamble functions were almost
identical except for the math applied to the coordinate. I opted to
combine them into one function that just looked at the snippet ID
to decide what expression to apply to the coordinate argument.

The default preamble generator is split out to have a helper
function that emits just a function declaration w/o any body so that
it can be used for all the other custom preamble generators. Also
generalized the string concatenation for both parameters and arguments
which will make it more readable as we extend what's passed in for
`append_defaults`.

The default expression generator is extended to support invoking
runtime effects (simply has to check for if a static function was
assigned or not).

Magic strings are also reduced in favor of modifying copies of
`kDefaultArgs`. The helper declaration always uses the parameters
with names matching `kDefaultArgs` so custom function bodies can
just refer to those variables as needed.

With the new preamble/expression consolidation, it was also feasible
to have ShaderNodes mask off SnippetRequirements for their input
variables that would be handled by composition or explicitly by
runtime effects. This makes it so that the generated helper functions
only declare the inputs that they actually are going to use.

Bug: b/349205749
Change-Id: I227666c6239f33ac62e95056a4ef4ae8cf4fa114
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/869619
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>