[graphite] Preliminary update to ShaderSnippet to support aggregate uniforms in structs
This adds the struct name field to ShaderSnippet and calculates the
required alignment at initialization time. For now runtime shader
snippets do not set a struct name so there's no logic to calculate
their required alignment (but that's planned).

The expression generation is updated to pass in the struct's field,
but no snippet uses structs yet and the top-level struct that is
generated doesn't know how to use structs yet, so this should not
result in any actual changes in behavior.

Bug: b/349208089
Change-Id: I53a43f3a7e84647b884071db214c7b8e9901e15b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/870877
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>