[graphite] Fix color text rendering, part 1.
We were treating the values read from the color atlas texture as
coverage, which is correct for LCD text but not pure color text.
We need to treat those values as color values, which will bring in
support for color filters. In addition, we need to disable
any shaders, as they shouldn't affect color text.

Bug: 347990502
Bug: 348683468
Change-Id: Ia1a5101ace10f0d1c798da4b5d978d992852511c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/871996
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>