[graphite] CoverBoundsRenderStep can be configured with custom DepthStencilSettings
This sets the ground work to have a Renderer that uses it for non-AA
fills (either paint with inverse fills, or the inner fills for convex
AA shapes).

This also removes the inverse field and instead bases the vertex bounds
choice based on the inverted property of the Shape itself. Shape
actually stores an extra bit to represent inverse primitive shapes
(not just paths), so this will fully address b/350603541.

Bug: b/350603541
Bug: b/333729316
Change-Id: I72f770fc5edd6ff93e3f9f2f405524c46dcb5814
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/872282
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>