Update XCode 11.4.1 to 15.4
When building with Xcode >= 12 and running on an iPhone 11 with iOS 13.6
CFBundleCopyResourceURL hangs when it should return NULL (the target
does not exist). This makes supporting iOS 13.6 difficult. Move to
iPhone 8 with iOS 13.3.1 which does not have this issue.

Xcode 15 and 16 have minimum supported deployment target of iOS 12.
Attempting to target anything older may not work. This necessitates
dropping support for building for iOS 11.

Bug: b/338243739
Bug: b/353704411
Change-Id: I8230b250ff38a4281705ad8eb3a75c003f6900b8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/873357
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>